Closes #2854.

```
CPython:
$ python3.7 -m timeit 'import pyarrow.pandas_compat as pdcompat'
1 loop, best of 5: 592 nsec per loop
$ python3.6 -m timeit 'import pyarrow.pandas_compat as pdcompat'
10 loops, best of 3: 0.302 usec per loop
$ python3.5 -m timeit 'import pyarrow.pandas_compat as pdcompat'
10 loops, best of 3: 0.946 usec per loop

Cython before:
$ python3.7 -m timeit -s 'from importbench import _import_bench' '_import_bench()'
1 loop, best of 5: 1.37 usec per loop

After change, without initialisation check:
$ python3.7 -m timeit -n 100000 -r 50 -s 'from importbench import _import_bench' '_import_bench()'
100000 loops, best of 50: 51.5 nsec per loop
$ python3.6 -m timeit -n 100000 -r 50 -s 'from importbench import _import_bench' '_import_bench()'
100000 loops, best of 50: 0.0424 usec per loop
$ python3.5 -m timeit -n 100000 -r 50 -s 'from importbench import _import_bench' '_import_bench()'
100000 loops, best of 50: 0.0454 usec per loop

After change, with initialisation check:
$ python3.7 -m timeit -r 20 -n 200000 -s 'from importbench import _import_bench' '_import_bench()'
200000 loops, best of 20: 119 nsec per loop
$ python3.6 -m timeit -r 20 -n 200000 -s 'from importbench import _import_bench' '_import_bench()'
200000 loops, best of 20: 0.0882 usec per loop
$ python3.5 -m timeit -r 20 -n 200000 -s 'from importbench import _import_bench' '_import_bench()'
200000 loops, best of 20: 0.107 usec per loop
```
Not sure what to do about Py2.7. It's not too bad in comparison, but I guess there are also some cases where calling `__import__()` (because that's what Cython does in Py2) isn't exactly the fastest.
```
$ python2.7 -m timeit -r 50 -n 200000 'import pyarrow.pandas_compat as pdcompat'
200000 loops, best of 50: 0.44 usec per loop
$ python2.7 -m timeit -r 50 -n 200000 -s 'from importbench import _import_bench' '_import_bench()'
200000 loops, best of 50: 0.691 usec per loop
```
Then again, its time is running up, so why bother…